% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_legend}
\alias{data_legend}
\title{Generate a source caption for plots}
\usage{
data_legend(data_df, struct_df = NULL)
}
\arguments{
\item{data_df}{A tibble returned by \code{\link[=get_data]{get_data()}}.}

\item{struct_df}{A tibble returned by \code{\link[=get_dimensions]{get_dimensions()}}. Optional.}
}
\value{
A character string suitable for plot captions.
}
\description{
Generate a source caption for plots
}
\examples{
\donttest{
if (trafa_available()) {
  d <- get_data("t10011", "itrfslut", ar = "2024")
  data_legend(d)
}}
}
