% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_minimize}
\alias{data_minimize}
\title{Remove monotonous columns from a data tibble}
\usage{
data_minimize(data_df)
}
\arguments{
\item{data_df}{A tibble returned by \code{\link[=get_data]{get_data()}}.}
}
\value{
A tibble with monotonous columns removed.
}
\description{
Remove monotonous columns from a data tibble
}
\examples{
\donttest{
if (trafa_available()) {
  d <- get_data("t10011", "itrfslut", ar = "2024")
  d |> data_minimize()
}}
}
