% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{dimension_extract_names}
\alias{dimension_extract_names}
\title{Extract dimension names}
\usage{
dimension_extract_names(dim_df)
}
\arguments{
\item{dim_df}{A tibble returned by \code{\link[=get_dimensions]{get_dimensions()}}.}
}
\value{
A character vector of dimension names.
}
\description{
Extract dimension names
}
\examples{
\donttest{
if (trafa_available()) {
  get_dimensions("t10011") |> dimension_extract_names()
}}
}
