% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{measure_extract_names}
\alias{measure_extract_names}
\title{Extract measure names}
\usage{
measure_extract_names(measure_df)
}
\arguments{
\item{measure_df}{A tibble returned by \code{\link[=get_measures]{get_measures()}}.}
}
\value{
A character vector of measure names.
}
\description{
Extract measure names
}
\examples{
\donttest{
if (trafa_available()) {
  get_measures("t10011") |> measure_extract_names()
}}
}
