% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products.R
\name{product_minimize}
\alias{product_minimize}
\title{Remove monotonous columns from a product tibble}
\usage{
product_minimize(product_df)
}
\arguments{
\item{product_df}{A tibble returned by \code{\link[=get_products]{get_products()}}.}
}
\value{
A tibble with monotonous columns removed.
}
\description{
Remove monotonous columns from a product tibble
}
\examples{
\donttest{
if (trafa_available()) {
  get_products() |> product_minimize()
}}
}
