% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTrafa-package.R
\docType{package}
\name{rTrafa-package}
\alias{rTrafa}
\alias{rTrafa-package}
\title{rTrafa: Access Swedish Transport Statistics from the 'Trafa' API}
\description{
A pipe-friendly R client for the Trafa statistical API (api.trafa.se). Provides a search-then-fetch workflow for discovering and downloading Swedish transport and infrastructure data using a consistent tibble-based interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lchansson.github.io/rTrafa/}
  \item \url{https://github.com/lchansson/rTrafa}
  \item Report bugs at \url{https://github.com/lchansson/rTrafa/issues}
}

}
\author{
\strong{Maintainer}: Love Hansson \email{love.hansson@gmail.com} [copyright holder]

}
\keyword{internal}
